ai/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0
flask>=2.3.0
flask-cors>=4.0.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0 